Fix snmp mac polling performance issue.
When SONiC system get a few hundreds MACs learnt on vlans, snmp-subagent
started to see below errors:
INFO supervisord: snmp-subagent socket.send() raised exception.
ERR snmp-subagent [ax_interface] ERROR: [Errno 32] Broken pipe

This was due to the snmpd getting timeout from subagent and closed the socket.

It was root caused to be the mac polling algorithm inefficiency.
The redis-server was taking average > 60% CPU cycles (peak to 85~90%) of one core.
After the fix in this commit, the redis-server load is average to
~3% (peak to ~6%) of one core.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>