The description of ifIndex in RFC1213 is
"A unique value for each interface.
Its valueranges between 1 and the value of ifNumber.
The value for each interface must remain constant at
least from one re-initialization of the entity's
network management system to the next reinitialization."

The based interface started from 0 is illegal in RFC1213 defined boundary.
So I think it should not minus one at this place.