[SNMP] Update description of entPhysicalDescr mib in case interface is not configured. (#205)
The ideas of this MIB is to reflect the physical entities and should not be affected by the config_db.json interface configuration.
In the case where the interface does not exist (not configured at all or as a result of user removing the interface from the config_db.json due to port breakout limitation on some platforms), entPhysicalDescr OID should return the transceiver type only and without the interface alias and the extra 'for ' text.

Signed-off-by: liora liora@nvidia.com

- What I did
In the case interface not configured but transceiver exist, the transceiver description is only the transceiver type without the extra 'for xxx' text.

- How I did it
Verify if interface is configured. If yes, add to the transceiver type the interface alias. Else, only the transceiver type.

- How to verify it
snmpwalk -v 2c -c public 10.210.24.85 1.3.6.1.2.1.47.1.1.1.1.2 | grep SFP

Signed-off-by: liora <liora@nvidia.com>