Added unit test to check #UC > half total queue behavior
Added a unit test subcase and relevant mock tablesto check for SNMP
behavior when there are more unicast queues than half of the total
number of queues.

This is to ensure SNMP does not operate on the false assumption where
there is always a 50-50 split between unicast and multicast queues.